Remove: index from skills
